backend/requirements.txt
fastapi==0.104.1
python-socketio==5.10.0
python-chess==1.999
uvicorn[standard]==0.24.0
aiofiles==23.2.1